Move AddressInput utils to separate files
