+++ apps/melding-form/src/app/(map)/locatie/kies/_components/AddressInput/AddressInput.tsx
@@ -22,7 +22,9 @@ import type { PDOKItem } from './types'
 import type { Coordinates } from '~/types'
 
 import { convertWktPointToCoordinates } from '../../utils'
-import { debounce, fetchAddressList, fetchAndSetAddress } from './utils'
+import { debounce } from './_utils/debounce'
+import { fetchAddressList } from './_utils/fetchAddressList'
+import { fetchAndSetAddress } from './_utils/fetchAndSetAddress'
 
 import styles from './AddressInput.module.css'
