--- apps/melding-form/src/app/(map)/locatie/kies/_components/AddressInput/utils.ts
@@ -1,77 +1,0 @@
-import type { useTranslations } from 'next-intl'
-
-import type { PDOKItem } from './types'
-import type { Coordinates } from '~/types'
-
-const pdokQueryParams =
-  'fq=bron:BAG&fq=type:adres&fq=gemeentenaam:(amsterdam "ouder-amstel" weesp)&fl=id,weergavenaam,centroide_ll&rows=7'
-
-// eslint-disable-next-line @typescript-eslint/no-unsafe-function-type
-export const debounce = (fn: Function, delay = 250) => {
-  let timer: ReturnType<typeof setTimeout>
-
-  return (...args: unknown[]) => {
-    clearTimeout(timer)
-    timer = setTimeout(() => fn(...args), delay)
-  }
-}
-
-export type AddressArgType = {
-  coordinates: Coordinates
-  setAddress: (address: string) => void
-  t: ReturnType<typeof useTranslations>
-}
-
-export const fetchAndSetAddress = async ({ coordinates: { lat, lng }, setAddress, t }: AddressArgType) => {
-  try {
-    const response = await fetch(
-      `https://api.pdok.nl/bzk/locatieserver/search/v3_1/reverse?lat=${lat}&lon=${lng}&rows=1&distance=30`,
-    )
-
-    if (!response.ok) {
-      throw new Error('Unable to fetch address from coordinates from PDOK')
-    }
-
-    const result = await response.json()
-    const address = result.response.docs?.[0]?.weergavenaam ?? t('no-address')
-
-    setAddress(address)
-  } catch (error) {
-    // eslint-disable-next-line no-console
-    console.error(error)
-    setAddress(t('no-address'))
-  }
-}
-
-export type AddressListArgType = {
-  setAddressList: (list: PDOKItem[]) => void
-  setShowListBox: (show: boolean) => void
-  value: string
-}
-
-export const fetchAddressList = async ({ setAddressList, setShowListBox, value }: AddressListArgType) => {
-  if (value.length < 3) {
-    setShowListBox(false)
-    setAddressList([])
-    return
-  }
-
-  try {
-    const response = await fetch(
-      `https://api.pdok.nl/bzk/locatieserver/search/v3_1/suggest?${pdokQueryParams}&q=${value}`,
-    )
-
-    if (!response.ok) {
-      throw new Error('Unable to fetch address suggestions from PDOK')
-    }
-
-    const result = await response.json()
-
-    setAddressList(result.response.docs)
-    setShowListBox(true)
-  } catch (error) {
-    // Only log the error, the user can continue without suggestions
-    // eslint-disable-next-line no-console
-    console.error(error)
-  }
-}
